+++ src/main/java/com/project/bankassetor/primary/model/response/AccountCreateResponse.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Builder
 @Getter
@@ -13,7 +15,7 @@ public class AccountCreateResponse {
     // 계좌 유형 (예: SAVING, CHECKING)
     private String accountType;
     private String accountNumber;
-    private int balance;
+    private BigDecimal balance;
 
     public static AccountCreateResponse of(Account account){
         return AccountCreateResponse.builder()
