+++ src/main/java/com/project/bankassetor/primary/model/response/TransactionHistoryResponse.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,8 +16,8 @@ import java.util.stream.Collectors;
 public class TransactionHistoryResponse {
 
     LocalDateTime transactionTime;  // 거래 시간
-    int transactionAmount;  // 거래 금액
-    int balanceAfter;  // 거래 후 잔액
+    BigDecimal transactionAmount;  // 거래 금액
+    BigDecimal balanceAfter;  // 거래 후 잔액
 
     public static List<TransactionHistoryResponse> of(List<CheckingTransactionHistory> transactionHistories) {
         return transactionHistories.stream()
